ISSUE #3905 - Removed duplicated code, uinfied sso hooks
