+++ frontend/src/v5/ui/components/authTemplate/authTemplate.styles.ts
@@ -24,9 +24,6 @@ import { Link } from 'react-router-dom';
 import { LOGIN_PATH } from '@/v5/ui/routes/routes.constants';
 import { Display } from '@/v5/ui/themes/media';
 import { Divider as DividerBase } from '@mui/material';
-import { Button } from '@controls/button';
-import MicrosoftIcon from '@assets/icons/thirdParty/microsoft.svg';
-import { createElement } from 'react';
 
 const customLogoPath = clientConfigService.getCustomLogoPath();
 export const customBackgroundPath = clientConfigService.getCustomBackgroundImagePath();
@@ -136,39 +133,4 @@ export const AuthSubHeader = styled.div`
 	color: ${({ theme }) => theme.palette.secondary.main};
 	user-select: none;
 	margin: 19px 0;
-`;
-
-export const MicrosoftButton = styled(Button).attrs({
-	component: Link,
-	variant: 'contained',
-	color: 'primary',
-	startIcon: createElement(MicrosoftIcon),
-})`
-	display: flex;
-	width: fit-content;
-	font-weight: 500;
-	font-size: 12px;
-	border-radius: 0;
-	background-color: #2F2F2F; /* The colour is hardcoded as this are microsoft specs and not part of the theme */
-
-	&:hover, &:active {
-		background-color: #2F2F2FF0; 
-
-	}
-
-	margin: 0;
-	padding:20px;
-`;
-
-export const NewSticker = styled.div`
-	color: ${({ theme }) => theme.palette.primary.main};
-	border: solid 1.5px ${({ theme }) => theme.palette.primary.main}; 
-	border-radius: 5px;
-	padding: 4px 6px;
-	display: inline;
-	font-size: 10px;
-	font-weight: 700;
-	top: -4px;
-	position: relative;
-	left: 8px;
 `;
