+++ frontend/src/v5/ui/routes/login/login.component.tsx
@@ -26,9 +26,11 @@ import { LoginSchema } from '@/v5/validation/userSchemes/loginSchemes';
 import { AuthTemplate } from '@components/authTemplate';
 import { AuthHooksSelectors } from '@/v5/services/selectorsHooks';
 import { SubmitButton } from '@controls/submitButton/submitButton.component';
-import { AuthSubHeader, Divider, MicrosoftButton, NewSticker } from '@components/authTemplate/authTemplate.styles';
+import { AuthSubHeader, Divider } from '@components/authTemplate/authTemplate.styles';
 import { useSSOLogin } from '@/v5/services/sso.hooks';
 import { Gap } from '@controls/gap';
+import { MicrosoftButton } from '@components/shared/sso/microsoftButton.component';
+import { NewSticker } from '@components/shared/sso/microsoftText.styles';
 import { AuthFormLogin, ForgotPasswordPrompt, OtherOptions, SignUpPrompt, UnhandledErrorInterceptor } from './login.styles';
 import { AuthHeading, ErrorMessage, FormPasswordField, FormUsernameField } from './components/components.styles';
 import { PASSWORD_FORGOT_PATH, RELEASE_NOTES_ROUTE, SIGN_UP_PATH } from '../routes.constants';
